Initial commit of KernelyzeBase files
KernelyzeBase is the foundational functionality; KernelyzeBaseTest is a
battery of tests for it.  KernelyzeBaseBindC offers C bindings to the
Fortran code in KernelyzeBase, and KernelyzeBaseBindCTest is a set of
tests of those bindings.  kernlcaml is a set of OCaml bindings via the C
bindings.  All projects were primarily developed on Windows (including
the OCaml functionality).  The project files in the repository currently
support Intel Fortran* (through Visual Studio*) and NAG Fortran*
(through the NAG Fortran Builder*).